requirements.txt
PyYAML
python-dotenv
mega.py
